C# Coding Standards: Export based on a few years of writing
The C# Coding Standards was started when writing software for .NET 4.0 and later. Since then, the .NET C# language has evolved. But to keep software accessible to others it is not always the best to use all new features. It is also important to provide rationale, so that others can take the coding standards and tailor them to their own needs.