CoreDump: Configuring Ubuntu
Describe my journey trying to capture core dumps on my C++ program when using sanitizers and why it didn't work. This lead to deep understanding of how Linux is configured to generate core dumps, the user space programs it calls, and then debug why for the Address Sanitizer, it still isn't being called.